Logger (#6)
* Initial Commit

* Configured Webpack for SCSS, CSS & JS

- Still need to setup reactjs

* Added support for React

Added support to webpack that allows it to handle reactJS

* Added a task Scheduler

* Fixed capitalization error

* Added unit tests

Added unit tests for the `Src/Services/Task` class. These tests make sure the methods will reject any value that would be out of bounds.

I've also restructured the project to meet the following layout. To accommodate this change the old src directory is now called resources
```
/
/Src
/Src/Resources
/Tests
```

* Added test support for Travis CI

* Added build script and updated travis CI script

* Added more branches to Travis CI

* Fixed Travis CI permission issue

* Updated readme

* Smaller Ymal script

* Added xenial distribution to the testing script

* Removed quote marks from YMAL

* Removed Travis CI testing

- Cannot get it to work, I've removed it for now.

* Created a basic file logger

* Clean up of code

Removed the test log code in the main method and removed the logic in Log.Info() as this will allways log.

* Clean logs daily 00:00 UTC

* Debug only works on development mode